requirements.txt
openai>=1.0.0
python-frontmatter>=1.0.0
pyyaml>=6.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.2
python-multipart>=0.0.20
python-docx>=1.1.2
pdfminer.six>=20231228
python-docx>=1.1.0
pdfminer.six>=20231228
python-docx>=1.1.2
pdfminer.six>=20240706